+++ packages/suite/src/views/dashboard/index.tsx
@@ -1,19 +1,20 @@
 import styled from 'styled-components';
 
 import { Context } from '@suite-common/message-system';
+import { selectHasBitcoinOnlyFirmware } from '@suite-common/wallet-core';
 import { Column } from '@trezor/components';
 import { spacings, spacingsPx } from '@trezor/theme';
 
 import { PageHeader } from 'src/components/suite/layouts/SuiteLayout';
 import { ContextMessage } from 'src/components/wallet/WalletLayout/AccountBanners/ContextMessage';
-import { useLayout } from 'src/hooks/suite';
+import { useLayout, useSelector } from 'src/hooks/suite';
 
 import { AssetsView } from './AssetsView/AssetsView';
 import { DashboardFooter } from './DashboardFooter';
 import { DashboardPassphraseBanner } from './DashboardPassphraseBanner';
 import { DashboardPromoBanner } from './DashboardPromoBanner/DashboardPromoBanner';
 import { PortfolioCard } from './PortfolioCard/PortfolioCard';
-import { StakeEthCard } from './StakeEthCard/StakeEthCard';
+import { StakingDashboard } from './StakingDashboard/StakingDashboard';
 import { useNotificationForDisconnectedDevice } from './useNotificationForDisconnectedDevice';
 
 const Container = styled.div`
@@ -26,6 +27,8 @@ export const Dashboard = () => {
     useLayout('Home', <PageHeader />);
     useNotificationForDisconnectedDevice();
 
+    const hasBitcoinOnlyFirmware = useSelector(selectHasBitcoinOnlyFirmware);
+
     return (
         <Column gap={spacings.xxxxl} data-testid="@dashboard/index">
             <Container>
@@ -35,7 +38,7 @@ export const Dashboard = () => {
             </Container>
             <DashboardPromoBanner />
             <AssetsView />
-            <StakeEthCard />
+            {!hasBitcoinOnlyFirmware && <StakingDashboard />}
             <DashboardFooter />
         </Column>
     );